Separated mneme integration from others
