+++ Mneme/Mneme.Desktop/Mneme.Desktop.csproj
@@ -50,6 +50,7 @@
     <ProjectReference Include="..\Mneme.Core\Mneme.Core.csproj" />
     <ProjectReference Include="..\Mneme.PrismModule.Configuration.Integration\Mneme.PrismModule.Configuration.Integration.csproj" />
     <ProjectReference Include="..\Mneme.PrismModule.Dashboard\Mneme.PrismModule.Dashboard.csproj" />
+    <ProjectReference Include="..\Mneme.PrismModule.Integration.Facade\Mneme.PrismModule.Integration.Facade.csproj" />
     <ProjectReference Include="..\Mneme.PrismModule.Integrations.GoogleBooks\Mneme.PrismModule.Integrations.GoogleBooks.csproj" />
     <ProjectReference Include="..\Mneme.PrismModule.Integrations.Mneme\Mneme.PrismModule.Integrations.Mneme.csproj" />
     <ProjectReference Include="..\Mneme.PrismModule.Integrations.Pluralsight\Mneme.PrismModule.Integrations.Pluralsight.csproj" />
