+++ Mneme/Mneme.Notes/MnemeNotesCreator.cs
@@ -1,4 +1,4 @@
-﻿using Mneme.Core.Interfaces;
+﻿using Mneme.Core;
 using Mneme.Integrations.Mneme.Contract;
 using Mneme.Model.Sources;
 
@@ -7,16 +7,18 @@ namespace Mneme.Notes
 	public class MnemeNotesCreator
 	{
 		private readonly IBundledIntegrationFacades integration;
+		private readonly IMnemeIntegrationFacade mnemeIntegration;
 
-		public MnemeNotesCreator(IBundledIntegrationFacades integration)
+		public MnemeNotesCreator(IBundledIntegrationFacades integration, IMnemeIntegrationFacade mnemeIntegration)
 		{
 			this.integration = integration;
+			this.mnemeIntegration = mnemeIntegration;
 		}
 		public async Task<MnemePreelaboration> SaveMnemeNote(SourcePreview sourcePreview, string content, string title, string path, CancellationToken ct)
 		{
 			var newSource = (await integration.GetSource(sourcePreview.Id, sourcePreview.TypeOfSource, ct)) as MnemeSource;
 			var note = new MnemePreelaboration() { IntegrationId = Guid.NewGuid().ToString(), Content = content, Title = title, Path = path, CreationTime = DateTime.Now, Source = newSource };
-			await integration.CreateNote(note, ct);
+			await mnemeIntegration.CreateNote(note, ct);
 			return note;
 		}
