+++ Mneme/Mneme.PrismModule.Integration.Facade/BundledIntegrationFacades.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Mneme.Core;
 using Mneme.Integrations.Contracts;
 using Mneme.Integrations.GoogleBooks.Contract;
 using Mneme.Integrations.Mneme.Contract;
@@ -10,7 +11,7 @@ using Mneme.Integrations.Pluralsight.Contract;
 using Mneme.Model.Preelaborations;
 using Mneme.Model.Sources;
 
-namespace Mneme.Core.Interfaces
+namespace Mneme.PrismModule.Integration.Facade
 {
 	public class BundledIntegrationFacades : IBundledIntegrationFacades
 	{
@@ -52,26 +53,7 @@ namespace Mneme.Core.Interfaces
 				await mnemeIntegration.UpdateSource(source, ct);
 			}
 		}
-		/// <summary>
-		/// TODO
-		/// </summary>
-		/// <param name="id"></param>
-		/// <returns></returns>
-		public async Task DeleteNote(string id, CancellationToken ct = default)
-		{
-			await mnemeIntegration.DeleteNote(id, ct);
-		}
-		/// <summary>
-		/// TODO
-		/// </summary>
-		/// <param name="id"></param>
-		/// <param name="type"></param>
-		/// <returns></returns>
-		/// <exception cref="NotImplementedException"></exception>
-		public async Task DeleteSource(string id, CancellationToken ct = default)
-		{
-			await mnemeIntegration.DeleteSource(id, ct);
-		}
+
 
 		public async Task<IReadOnlyList<Preelaboration>> GetActiveNotes(CancellationToken ct = default)
 		{
@@ -119,9 +101,7 @@ namespace Mneme.Core.Interfaces
 			ret.AddRange(await pluralsightIntegration.GetKnownSources(onlyActive, ct));
 			return ret;
 		}
-		/// <summary>
-		/// Only Mneme
-		/// </summary>
+
 		public async Task IgnoreSource(string id, string type, CancellationToken ct = default)
 		{
 			if (type == GoogleBooksSource.Type)
@@ -151,43 +131,11 @@ namespace Mneme.Core.Interfaces
 			await pluralsightIntegration.MigrateDatabase(ct);
 			//await testingModule.MigrateDatabase(ct); //TODO
 		}
-		/// <summary>
-		/// Only Mneme
-		/// </summary>
-		public async Task UpdateSource(string id, string title, string details, CancellationToken ct = default)
-		{
-			var source = await mnemeIntegration.GetSource(id, ct);
-			source.Title = title;
-			source.Details = details;
-			await mnemeIntegration.UpdateSource(source, ct);
-		}
-
-		public async Task<bool> Create(MnemeSource source, CancellationToken ct = default)
-		{
-			//work around to problem where integration id's were different but for some reason db was telling it already exists
-			try
-			{
-				var sourceDb = await mnemeIntegration.GetSource(source.IntegrationId, ct);
-				return false;
-			}
-			catch (InvalidOperationException) 
-			{
-				await mnemeIntegration.CreateSource(source);
-				return true;
-			}
-		}
-
-		public async Task CreateNote(MnemePreelaboration note, CancellationToken ct = default)
-		{
-			await mnemeIntegration.CreateNote(note);
-		}
 
 		public async Task<Source> GetSource(string id, string type, CancellationToken ct)
 		{
 			if (type == GoogleBooksSource.Type)
-			{
 				return await googleBooksIntegration.GetSource(id, ct);
-			}
 			else if (type == MnemeSource.Type)
 			{
 				return await mnemeIntegration.GetSource(id, ct);
