+++ Mneme/Mneme.Sources/MnemeSourceManager.cs
@@ -1,4 +1,4 @@
-﻿using Mneme.Core.Interfaces;
+﻿using Mneme.Core;
 using Mneme.Integrations.Mneme.Contract;
 
 namespace Mneme.Sources
@@ -6,22 +6,24 @@ namespace Mneme.Sources
 	public class MnemeSourceManager : IDisposable
 	{
 		private readonly IBundledIntegrationFacades integration;
+		private readonly IMnemeIntegrationFacade mnemeIntegration;
 
-		public MnemeSourceManager(IBundledIntegrationFacades integration)
+		public MnemeSourceManager(IBundledIntegrationFacades integration, IMnemeIntegrationFacade mnemeIntegration)
 		{
 			this.integration = integration;
+			this.mnemeIntegration = mnemeIntegration;
 		}
 
 		public async Task<MnemeSource?> SaveMnemeSource(string sourceTitle, string details, CancellationToken ct)
 		{
 			var source = new MnemeSource { Title = sourceTitle, Details = details, Active = true, IntegrationId = MnemeSource.GenerateIntegrationId(sourceTitle, details) };
-			var success = await integration.Create(source, ct);
+			var success = await mnemeIntegration.Create(source, ct);
 			return success ? source : null;
 		}
 
 		public async Task <MnemeSource?> UpdateMnemeSource(string id, string title, string details, CancellationToken ct)
 		{
-			await integration.UpdateSource(id, title, details);
+			await mnemeIntegration.UpdateSource(id, title, details);
 			return (MnemeSource)await integration.GetSource(id, MnemeSource.Type);
 		}
