+++ Mneme/Mneme.sln
@@ -72,7 +72,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Testing", "Testing", "{20BA
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mneme.PrismModule.Testing", "Mneme.PrismModule.Testing\Mneme.PrismModule.Testing.csproj", "{B486CE90-EEA0-4C60-B79B-E73F93FF8FAE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mneme.PrismModule.Configuration.Integration", "Mneme.PrismModule.Configuration.Integration\Mneme.PrismModule.Configuration.Integration.csproj", "{BFC16924-63A3-4294-BCC7-47C9EF7F742E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mneme.PrismModule.Configuration.Integration", "Mneme.PrismModule.Configuration.Integration\Mneme.PrismModule.Configuration.Integration.csproj", "{BFC16924-63A3-4294-BCC7-47C9EF7F742E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mneme.PrismModule.Integration.Facade", "Mneme.PrismModule.Integration.Facade\Mneme.PrismModule.Integration.Facade.csproj", "{9680235C-EDC0-47DF-AF74-950E5386BCA7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -176,6 +178,10 @@ Global
 		{BFC16924-63A3-4294-BCC7-47C9EF7F742E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BFC16924-63A3-4294-BCC7-47C9EF7F742E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BFC16924-63A3-4294-BCC7-47C9EF7F742E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9680235C-EDC0-47DF-AF74-950E5386BCA7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9680235C-EDC0-47DF-AF74-950E5386BCA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9680235C-EDC0-47DF-AF74-950E5386BCA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9680235C-EDC0-47DF-AF74-950E5386BCA7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -213,6 +219,7 @@ Global
 		{20BA2992-8A08-46DA-9565-1B6AE0B83D99} = {1950FCDB-AEB2-4448-BE2B-92837D26C16B}
 		{B486CE90-EEA0-4C60-B79B-E73F93FF8FAE} = {20BA2992-8A08-46DA-9565-1B6AE0B83D99}
 		{BFC16924-63A3-4294-BCC7-47C9EF7F742E} = {1950FCDB-AEB2-4448-BE2B-92837D26C16B}
+		{9680235C-EDC0-47DF-AF74-950E5386BCA7} = {AEA51CCF-4A41-489B-B69F-661EB884F5A7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1D9D3402-54B7-474F-A18E-D8F86A41C1FA}
